requirements.txt
aiogram>=3.0
python-dotenv
openpyxl
aiosqlite
apscheduler